Make GroovyScriptRunner#invoke() return CompletableFuture
This allows Groovy scripts to invoke other scripts asynchronously using the Java 8 CompletableFuture
API. With this change, if the script wants to return a value to the caller, it must invoke the
result() method, passing in the value to return. When relying on the result of another script
invocation, you would call this from inside the thenApply() block.